fix: fix error change relative path to absolute path and fix circular import error #83
