+++ optimizers/optims.py
@@ -1,5 +1,16 @@
-from ._util import optimizer_entrypoint, is_optimizer
+from torch.optim import SGD, Adagrad, Adam
 
+OPTIMIZER_ENTRYPOINTS = {
+    'sgd': SGD,
+    'adagrad': Adagrad,
+    'adam': Adam,
+}
+
+def optimizer_entrypoint(optimizer_name):
+    return OPTIMIZER_ENTRYPOINTS[optimizer_name]
+
+def is_optimizer(optimizer_name):
+    return optimizer_name in OPTIMIZER_ENTRYPOINTS
 
 def create_optimizer(optimizer_name, parameter, args):
     """_summary_
@@ -32,4 +43,3 @@ def create_optimizer(optimizer_name, parameter, args):
     else:
         raise RuntimeError('Unknown optimizer (%s)' % optimizer_name)
     return optimizer
-
