+++ train.py
@@ -31,19 +31,19 @@ def train(data_dir, model_dir, args):
 
     for k,(train_loader, val_loader) in enumerate(zip(t_loaders,v_loaders)):
         # -- model
-        model_module = getattr(import_module("models/my_model"), args.model)  # default: BaseModel
+        model_module = getattr(import_module("models.my_model"), args.model)  # default: BaseModel
         model = model_module(
             num_classes=args.num_classes
         )
         if args.canny:
             backbone = model
-            model_module = getattr(import_module("models/my_model"), "Canny")
+            model_module = getattr(import_module("models.my_model"), "Canny")
             model = model_module(
                 backbone = backbone,
             )
         elif args.arcface:
             backbone = model_module(num_classes=1000)
-            model_module = getattr(import_module("models/my_model"), "ArcfaceModel")
+            model_module = getattr(import_module("models.my_model"), "ArcfaceModel")
             model = model_module(
                 backbone = backbone,
                 num_features = 1000,
